Move displaced test classes
