+++ hazelcast-jet-core/src/test/java/com/hazelcast/jet/impl/deployment/ApacheV1andV2.java
@@ -14,27 +14,39 @@
  * limitations under the License.
  */
 
-package deployment;
+package com.hazelcast.jet.impl.deployment;
 
 import com.hazelcast.jet.AbstractProcessor;
+import java.io.BufferedReader;
+import java.io.IOException;
+import java.net.URISyntaxException;
+import java.net.URL;
+import java.nio.file.Files;
+import java.nio.file.Paths;
 
+import static org.junit.Assert.assertNotNull;
+import static org.junit.Assert.assertTrue;
 import static org.junit.Assert.fail;
 
-public class LoadPersonAndCar extends AbstractProcessor {
+public class ApacheV1andV2 extends AbstractProcessor {
 
     @Override
     public boolean complete() {
         ClassLoader cl = getClass().getClassLoader();
+        URL resource = cl.getResource("apachev1");
+        assertNotNull(resource);
+        BufferedReader reader = null;
         try {
-            cl.loadClass("com.sample.pojo.car.Car");
-        } catch (ClassNotFoundException e) {
-            fail();
-        }
-        try {
-            cl.loadClass("com.sample.pojo.person.Person$Appereance");
-        } catch (ClassNotFoundException ignored) {
+            reader = Files.newBufferedReader(Paths.get(resource.toURI()));
+            String firstLine = reader.readLine();
+            String secondLine = reader.readLine();
+            assertTrue(secondLine.contains("Version 1.1"));
+            assertNotNull(cl.getResourceAsStream("apachev2"));
+
+        } catch (IOException | URISyntaxException e) {
             fail();
         }
         return true;
+
     }
 }
