+++ hazelcast-jet-core/src/test/java/com/hazelcast/jet/impl/deployment/ApacheV2.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package deployment;
+package com.hazelcast.jet.impl.deployment;
 
 import com.hazelcast.jet.AbstractProcessor;
 import java.io.BufferedReader;
